Move dispatcher to top-level pkg.
Also, move pkg/heartbeat to dispatcher/heartbeat and get rid of pkg/.
